requirements.txt
fastf1
pandas
matplotlib
google-genai
fastapi
uvicorn[standard]
requests
